requirements.txt
appdirs==1.4.4
asgiref==3.2.10
bubbles==0.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.1
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
numpy==1.19.2
pandas==1.1.3
Pillow==7.2.0
pycrypto==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
virtualenv==20.0.35
